+++ src/dsp/oscillator.c
@@ -1,14 +1,14 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include "fix16.h"
-#include "oscillator.h"
+#include "dsp/oscillator.h"
 
 void oscIncreasePhase(Oscillator* oscillator) {
     oscillator->phase += oscillator->phaseIncrement;
     if (oscillator->phase >= oscillator->tableLen) oscillator->phase -= oscillator->tableLen;
 }
 
-void oscInit(Oscillator* osc, int16_t *table, size_t tableLen, float freq, float mod_index, float sample_rate) {
+void oscInit(Oscillator* osc, float *table, size_t tableLen, float freq, float mod_index, float sample_rate) {
     osc->table = table;
     osc->tableLen = tableLen;
     osc->phase = 0.0f;
